+++ CodeRebirthLib/src/Data/IWeighted.cs
@@ -1,0 +1,6 @@
+namespace CodeRebirthLib.Data;
+
+public interface IWeighted
+{
+    int Weight { get; }
+}